@@ -2,39 +2,54 @@
     "name": "@markusjx/java",
     "version": "1.0.1",
     "description": "A bridge between node.js and java",
-    "main": "index.js",
-    "typings": "index.d.ts",
+    "main": "dist/index.js",
+    "typings": "dist/index.d.ts",
     "repository": "https://github.com/MarkusJx/node-java-bridge",
     "engines": {
-        "node": ">=10"
+        "node": ">=14"
+    },
+    "bin": {
+        "java-ts-gen": "./dist/scripts/cli.js"
     },
     "scripts": {
-        "test": "mocha -r ts-node/register test/**/*.ts",
+        "test": "mocha -r ts-node/register test/**/*.test.ts",
         "pretest": "npm run build",
-        "test:debug": "mocha -r ts-node/register test/**/*.ts --debug",
+        "test:debug": "mocha -r ts-node/register test/**/*.test.ts --debug",
         "pretest:debug": "npm run build:dev",
-        "build": "cmake-js compile",
+        "build": "cmake-js build",
         "build:dev": "cmake-js compile -D --CDLOGGING=ON",
-        "pretest:dev": "npm run build:dev",
-        "test:dev": "node test.js --debug",
-        "postinstall": "node --unhandled-rejections=strict scripts/setJavaDllPath.js && cmake-js build",
-        "clean": "cmake-js clean"
+        "prebuild": "npm run build:ts",
+        "prebuild:dev": "npm run prebuild",
+        "build:ts": "tsc && cpy \"ts-src/**/*.d.ts\" dist/",
+        "postinstall": "ts-node ts-src/scripts/setJavaDllPath.ts && npm run build",
+        "clean": "cmake-js clean",
+        "benchmark": "ts-node -P test/tsconfig.json test/benchmark/benchmark.ts",
+        "prebenchmark": "npm run build",
+        "prettier:fix": "prettier --write ."
     },
     "author": "MarkusJx",
     "license": "MIT",
     "dependencies": {
         "@markusjx/n-api-tools": "^1.2.0",
+        "chalk": "^5.0.1",
         "cmake-js": "^6.3.2",
+        "cpy-cli": "^4.1.0",
         "find-java-home": "^1.2.2",
-        "node-addon-api": "^5.0.0"
+        "node-addon-api": "^5.0.0",
+        "ora": "^6.1.2",
+        "typescript": "^4.7.4",
+        "yargs": "^17.5.1"
     },
     "devDependencies": {
         "@types/bindings": "^1.5.1",
+        "@types/java": "^0.9.2",
         "@types/mocha": "^9.0.0",
-        "@types/node": "^18.0.0",
+        "@types/node": "^18.0.3",
+        "@types/yargs": "^17.0.10",
         "expose-gc": "^1.0.0",
         "mocha": "^10.0.0",
-        "ts-node": "^10.8.1",
-        "typescript": "^4.7.4"
+        "nanobench": "^2.1.1",
+        "prettier": "^2.7.1",
+        "ts-node": "^10.8.2"
     }
 }
