@@ -1,10 +1,8 @@
-import java, {java_instance_proxy} from "../index";
-import assert = require("assert");
+import java, { JavaClassInstance } from '../.';
+import assert = require('assert');
 import LogLevel = java.logging.LogLevel;
-import * as fs from "fs";
-import path from "path";
 
-declare class JString extends java_instance_proxy {
+declare class JString extends JavaClassInstance {
     static valueOf(values: string[]): Promise<JString>;
 
     static valueOfSync(values: string[]): JString;
@@ -21,7 +19,7 @@ declare class JString extends java_instance_proxy {
 
     toCharArray(): Promise<string[]>;
 
-    getBytesSync(): number[];
+    getBytesSync(): Buffer;
 
     splitSync(regex: string): string[];
 }
@@ -34,7 +32,7 @@ describe('StringTest', () => {
     });
 
     it('Check dll path', async () => {
-        const data = require("../jvmLibPath.json");
+        const data = require('../jvmLibPath.json');
         assert.strictEqual(await java.findJVM(), data);
     });
 
@@ -45,38 +43,38 @@ describe('StringTest', () => {
 
     it('Cached import', () => {
         java.importClass('java.lang.String');
-    })
+    });
 
     it('Async class resolve', async () => {
-        const JavaStringAsync = await java.importClassAsync('java.lang.String') as typeof JString;
-        const instance = await JavaStringAsync.newInstance('some text') as JString;
+        const JavaStringAsync = (await java.importClassAsync('java.lang.String')) as typeof JString;
+        const instance = (await JavaStringAsync.newInstance('some text')) as JString;
 
-        assert.strictEqual(await instance.toString(), "some text");
+        assert.strictEqual(await instance.toString(), 'some text');
     });
 
     let s1: JString;
     it('Create string instance', () => {
-        s1 = new JavaString("some text");
-        assert.strictEqual(s1.toStringSync(), "some text");
+        s1 = new JavaString('some text');
+        assert.strictEqual(s1.toStringSync(), 'some text');
     });
 
     it('String async match', async function () {
-        assert.strictEqual(await s1.toString(), "some text");
+        assert.strictEqual(await s1.toString(), 'some text');
     });
 
     it('String equals', () => {
-        assert.strictEqual(new JavaString("some text").equalsSync(s1), true);
-        assert.strictEqual(new JavaString("anything").equalsSync(s1), false);
+        assert.strictEqual(new JavaString('some text').equalsSync(s1), true);
+        assert.strictEqual(new JavaString('anything').equalsSync(s1), false);
     });
 
     it('String async equals', async () => {
-        let s2: JString = new JavaString("some text");
+        let s2: JString = new JavaString('some text');
         assert.strictEqual(await s2.equals(s1), true);
-        assert.strictEqual(await s2.equals(new JavaString("abc")), false);
+        assert.strictEqual(await s2.equals(new JavaString('abc')), false);
     });
 
     it('String async create', async function () {
-        let s2 = await JavaString.newInstance("some text") as JString;
+        let s2 = (await JavaString.newInstance('some text')) as JString;
         assert.strictEqual(await s2.equals(s1), true);
     });
 
@@ -87,23 +85,24 @@ describe('StringTest', () => {
 
     it('String to char array', () => {
         let arr = s1.toCharArraySync();
-        assert.strictEqual(JSON.stringify(arr), JSON.stringify(s1.toStringSync().split("")));
+        assert.strictEqual(JSON.stringify(arr), JSON.stringify(s1.toStringSync().split('')));
     });
 
     it('String to char array async', async () => {
         let arr = await s1.toCharArray();
-        assert.strictEqual(JSON.stringify(arr), JSON.stringify(s1.toStringSync().split("")));
-    })
+        assert.strictEqual(JSON.stringify(arr), JSON.stringify(s1.toStringSync().split('')));
+    });
 
     it('String to byte array', async () => {
         let arr = s1.getBytesSync();
-        let bytes = [...Buffer.from("some text")];
+        // @ts-ignore
+        let bytes = Buffer.from('some text');
 
         assert.strictEqual(JSON.stringify(arr), JSON.stringify(bytes));
     });
 
     it('String split', () => {
-        const split = s1.splitSync(" ");
-        assert.strictEqual(JSON.stringify(split), JSON.stringify(["some", "text"]));
+        const split = s1.splitSync(' ');
+        assert.strictEqual(JSON.stringify(split), JSON.stringify(['some', 'text']));
     });
 });
