@@ -1,10 +1,10 @@
-import { JavaType, java_instance_proxy } from "../index";
+import { JavaType, JavaClassInstance } from '../.';
 
 /**
  * A java collection
  * @see https://docs.oracle.com/javase/8/docs/api/java/util/Collection.html
  */
-export default class Collection<E extends JavaType> extends java_instance_proxy {
+export default class Collection<E extends JavaType> extends JavaClassInstance {
     /**
      * Ensures that this collection contains the specified element (optional operation).
      * Returns true if this collection changed as a result of the call.
@@ -13,14 +13,14 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * be added to this collection. In particular, some collections will refuse to add null elements,
      * and others will impose restrictions on the type of elements that may be added. Collection
      * classes should clearly specify in their documentation any restrictions on what elements may be added.
-     * 
+     *
      * If a collection refuses to add a particular element for any reason other than that
      * it already contains the element, it must throw an exception (rather than returning false).
      * This preserves the invariant that a collection always contains the specified element after
      * this call returns.
-     * 
+     *
      * Async call.
-     * 
+     *
      * @param e element whose presence in this collection is to be ensured
      * @returns true if this collection changed as a result of the call
      */
@@ -34,14 +34,14 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * be added to this collection. In particular, some collections will refuse to add null elements,
      * and others will impose restrictions on the type of elements that may be added. Collection
      * classes should clearly specify in their documentation any restrictions on what elements may be added.
-     * 
+     *
      * If a collection refuses to add a particular element for any reason other than that
      * it already contains the element, it must throw an exception (rather than returning false).
      * This preserves the invariant that a collection always contains the specified element after
      * this call returns.
-     * 
+     *
      * Sync call.
-     * 
+     *
      * @param e element whose presence in this collection is to be ensured
      * @returns true if this collection changed as a result of the call
      */
@@ -53,7 +53,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * the operation is in progress. (This implies that the behavior of this call is undefined
      * if the specified collection is this collection, and this collection is nonempty.)
      * Async call.
-     * 
+     *
      * @param c collection containing elements to be added to this collection
      * @returns true if this collection changed as a result of the call
      */
@@ -65,7 +65,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * the operation is in progress. (This implies that the behavior of this call is undefined
      * if the specified collection is this collection, and this collection is nonempty.)
      * Sync call.
-     * 
+     *
      * @param c collection containing elements to be added to this collection
      * @returns true if this collection changed as a result of the call
      */
@@ -90,7 +90,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * More formally, returns true if and only if this collection contains at
      * least one element e such that (o==null ? e==null : o.equals(e)).
      * Async call.
-     * 
+     *
      * @param o element whose presence in this collection is to be tested
      * @returns true if this collection contains the specified element
      */
@@ -101,7 +101,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * More formally, returns true if and only if this collection contains at
      * least one element e such that (o==null ? e==null : o.equals(e)).
      * Sync call.
-     * 
+     *
      * @param o element whose presence in this collection is to be tested
      * @returns true if this collection contains the specified element
      */
@@ -110,7 +110,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
     /**
      * Returns true if this collection contains all of the elements in the specified collection.
      * Async call.
-     * 
+     *
      * @param c collection to be checked for containment in this collection
      * @returns true if this collection contains all of the elements in the specified collection
      */
@@ -119,7 +119,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
     /**
      * Returns true if this collection contains all of the elements in the specified collection.
      * Sync call.
-     * 
+     *
      * @param c collection to be checked for containment in this collection
      * @returns true if this collection contains all of the elements in the specified collection
      */
@@ -128,7 +128,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
     /**
      * Returns true if this collection contains no elements.
      * Async call.
-     * 
+     *
      * @returns true if this collection contains no elements
      */
     isEmpty(): Promise<boolean>;
@@ -136,7 +136,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
     /**
      * Returns true if this collection contains no elements.
      * Sync call.
-     * 
+     *
      * @returns true if this collection contains no elements
      */
     isEmptySync(): boolean;
@@ -148,7 +148,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * Returns true if this collection contained the specified element
      * (or equivalently, if this collection changed as a result of the call).
      * Async call.
-     * 
+     *
      * @param o element to be removed from this collection, if present
      * @returns true if an element was removed as a result of this call
      */
@@ -161,7 +161,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * Returns true if this collection contained the specified element
      * (or equivalently, if this collection changed as a result of the call).
      * Sync call.
-     * 
+     *
      * @param o element to be removed from this collection, if present
      * @returns true if an element was removed as a result of this call
      */
@@ -172,7 +172,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * the specified collection (optional operation). After this call returns,
      * this collection will contain no elements in common with the specified collection.
      * Async call.
-     * 
+     *
      * @param c collection containing elements to be removed from this collection
      * @returns true if this collection changed as a result of the call
      */
@@ -183,7 +183,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * the specified collection (optional operation). After this call returns,
      * this collection will contain no elements in common with the specified collection.
      * Sync call.
-     * 
+     *
      * @param c collection containing elements to be removed from this collection
      * @returns true if this collection changed as a result of the call
      */
@@ -195,7 +195,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * removes from this collection all of its elements that are not
      * contained in the specified collection.
      * Async call.
-     * 
+     *
      * @param c collection containing elements to be retained in this collection
      * @returns true if this collection changed as a result of the call
      */
@@ -207,7 +207,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * removes from this collection all of its elements that are not
      * contained in the specified collection.
      * Sync call.
-     * 
+     *
      * @param c collection containing elements to be retained in this collection
      * @returns true if this collection changed as a result of the call
      */
@@ -218,19 +218,19 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * If this collection contains more than Integer.MAX_VALUE elements,
      * returns Integer.MAX_VALUE.
      * Async call.
-     * 
+     *
      * @returns the number of elements in this collection
      */
     size(): Promise<number>;
 
     /**
-    * Returns the number of elements in this collection.
-    * If this collection contains more than Integer.MAX_VALUE elements,
-    * returns Integer.MAX_VALUE.
-    * Async call.
-    * 
-    * @returns the number of elements in this collection
-    */
+     * Returns the number of elements in this collection.
+     * If this collection contains more than Integer.MAX_VALUE elements,
+     * returns Integer.MAX_VALUE.
+     * Async call.
+     *
+     * @returns the number of elements in this collection
+     */
     sizeSync(): number;
 
     /**
@@ -238,7 +238,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * If this collection makes any guarantees as to what order its elements
      * are returned by its iterator, this method must return the elements in the same order.
      * Async call.
-     * 
+     *
      * @returns an array containing all of the elements in this collection
      */
     toArray(): Promise<E[]>;
@@ -248,7 +248,7 @@ export default class Collection<E extends JavaType> extends java_instance_proxy 
      * If this collection makes any guarantees as to what order its elements
      * are returned by its iterator, this method must return the elements in the same order.
      * Sync call.
-     * 
+     *
      * @returns an array containing all of the elements in this collection
      */
     toArraySync(): E[];
